#23 - Delete files that do not need to share search components
