+++ front/src/containers/TopLayout.tsx
@@ -6,9 +6,10 @@ import Tab from '@material-ui/core/Tab';
 import style from "styled-components"
 
 import * as Styles from "./styles";
-import TrackContainer from "./track/TrackCotainer";
-import AlbumContainer from "./album/AlbumContainer";
-import ArtistContainer from "./artist/ArtistContainer";
+import AlbumLayoutComponent from "./album/AlbumLayoutComponent";
+import ArtistLayoutComponent from "./artist/ArtistLayoutComponent";
+import TrackLayoutComponent from "./track/TrackLayoutComponent";
+import SearchLayout from "./AllSearchComponent";
 import "./top.css";
 
 const Div = style.div`
@@ -26,7 +27,7 @@ const TopLayout: React.FC<Props> = ({
   const handleChange = (event: React.ChangeEvent<{}>, newValue: number) => {
     setValue(newValue);
   };
-
+  
   return (
     <>
       <Div>
@@ -43,14 +44,15 @@ const TopLayout: React.FC<Props> = ({
                 <Tab label="Album" {...Styles.a11yProps(2)} />
               </Tabs>
             </AppBar>
+            <SearchLayout token={token} />
             <Styles.TabPanel  value={value} index={0}>
-              <TrackContainer token={token} />
+              <TrackLayoutComponent token={token} />
             </Styles.TabPanel>
             <Styles.TabPanel value={value} index={1}>
-              <ArtistContainer token={token} />
+              <ArtistLayoutComponent token={token} />
             </Styles.TabPanel>
             <Styles.TabPanel value={value} index={2}>
-              <AlbumContainer token={token} />
+              <AlbumLayoutComponent token={token} />
             </Styles.TabPanel>
           </div>
         </div>
