+++ front/src/containers/styles.tsx
@@ -39,6 +39,6 @@ export function a11yProps(index: any) {
 export const useStyles = makeStyles((theme: Theme) => ({
   root: {
     flexGrow: 1,
-    backgroundColor: theme.palette.background.paper,
+    backgroundColor: 'black',
   },
 }));
