+++ front/src/containers/track/track.css
@@ -1,19 +1,5 @@
-.track {
-	height: auto;
-	padding: 10px 8px;
-}
-.track:hover {
-	background-color: #101010;
-}
-.counter {
-	float: left;
-	width: 24px;
-}
-.counter p {
-	font-size: 12px;
-	color: #ccc;
-	text-align: right;
-	padding-top: 4px;
+.track-wrapper {
+	padding: 16px;
 }
 .track img {
 	background-repeat: no-repeat;
@@ -28,38 +14,32 @@
 	-moz-transition: opacity 0.15s ease-in-out;
 	transition: opacity 0.15s ease-in-out;
 }
-.track-artists {
-	padding-left: 40px;
+.track {
+	width: 280px;
+	vertical-align: top;
+	margin-left: auto;
+	margin-right: auto;
+	text-align: center;
 }
-.track-artists h5 {
-	font-size: 16px;
-	font-weight: 100;
-}
-.track-artists li {
-	display: inline-block;
-	padding-right: 8px;
-	font-size: 13px;
-	color: #ccc;
+.track-wrapper p {
+	padding-top: 6px;
+	padding-bottom: 32px;
 }
 @media screen and (min-width: 620px) {
+	.track-wrapper {
+		text-align: center;
+	}
 	.track {
-		margin-right: 32px;
+		width: 260px;
 	}
-	.tracks-wrapper {
-		padding-left: 32px;
-	}
-	.counter p {
-		font-size: 14px;
-	}
-	.track-artists h5 {
-		font-size: 18px;
-	}
+
 	.track img {
 		width: 240px;
 		height: 240px;
 	}
-	.track-artists li {
-		font-size: 15px;
-		padding-top: 2px;
+
+	.track {
+		display: inline-block;
+		padding: 0px 12px;
 	}
 }
